# Reduce multiple consecutive values in each thread to improve efficiency