--- a/src/mapreduce.jl
+++ b/src/mapreduce.jl
@@ -91,11 +91,11 @@ Base.@propagate_inbounds _map_getindex(args::Tuple{}, I) = ()
 # product of the two iterators `Rreduce` and `Rother`, where the latter iterator will have
 # singleton entries for the dimensions that should be reduced (and vice versa).
 function partial_mapreduce_device(f, op, neutral, maxthreads, ::Val{Rreduce},
-    ::Val{Rother}, ::Val{Rlen}, ::Val{stride}, shuffle, R, As...) where {Rreduce, Rother, Rlen, stride}
+    ::Val{Rother}, ::Val{Rlen}, ::Val{grain}, shuffle, R, As...) where {Rreduce, Rother, Rlen, grain}
     # decompose the 1D hardware indices into separate ones for reduction (across items
     # and possibly groups if it doesn't fit) and other elements (remaining groups)
     localIdx_reduce = thread_position_in_threadgroup_1d()
-    localDim_reduce = threads_per_threadgroup_1d() * stride
+    localDim_reduce = threads_per_threadgroup_1d() * grain
     groupIdx_reduce, groupIdx_other = fldmod1(threadgroup_position_in_grid_1d(), Rlen)
 
     # group-based indexing into the values outside of the reduction dimension
@@ -115,8 +115,8 @@ function partial_mapreduce_device(f, op, neutral, maxthreads, ::Val{Rreduce},
         val = op(neutral, neutral)
 
         # read multiple consecutive values in reduction dimension to improve efficiency
-        ireduce = (localIdx_reduce - 1) * stride + (groupIdx_reduce - 1) * localDim_reduce
-        limit = ireduce + stride
+        ireduce = (localIdx_reduce - 1) * grain + (groupIdx_reduce - 1) * localDim_reduce
+        limit = ireduce + grain
         while ireduce < limit
             ireduce += 1
             next = if ireduce <= length(Rreduce)
@@ -171,9 +171,9 @@ function GPUArrays.mapreducedim!(f::F, op::OP, R::WrappedMtlArray{T},
     # but allows us to write a generalized kernel supporting partial reductions.
     R′ = reshape(R, (size(R)..., 1))
 
-    # number of consecutive values in reduction dimension read by each thread
-    # TODO: experiment and document this choice
-    stride = 16 ÷ sizeof(T)
+    # number of consecutive values in reduction dimension read by each thread.
+    # based on experiments, 4 is a good value for all scalar types.
+    grain = 4
 
     # how many threads can we launch?
     #
@@ -182,7 +182,7 @@ function GPUArrays.mapreducedim!(f::F, op::OP, R::WrappedMtlArray{T},
     # so that we can span the entire reduction dimension using a single item group.
     # XXX: can we query the 1024?
     kernel = @metal launch=false partial_mapreduce_device(f, op, init, Val(1024), Val(Rreduce), Val(Rother),
-                                                          Val(UInt64(length(Rother))), Val(stride), Val(shuffle), R′, A)
+                                                          Val(UInt64(length(Rother))), Val(grain), Val(shuffle), R′, A)
     pipeline = MTLComputePipelineState(kernel.fun.device, kernel.fun)
 
     # how many threads do we want?
@@ -206,7 +206,7 @@ function GPUArrays.mapreducedim!(f::F, op::OP, R::WrappedMtlArray{T},
     # optimal as it might not saturate the GPU. we already launch some groups to process
     # independent dimensions in parallel; pad that number to ensure full occupancy.
     other_groups = length(Rother)
-    reduce_groups = cld(length(Rreduce), reduce_threads * stride)
+    reduce_groups = cld(length(Rreduce), reduce_threads * grain)
 
     # determine the launch configuration
     threads = reduce_threads
@@ -217,7 +217,7 @@ function GPUArrays.mapreducedim!(f::F, op::OP, R::WrappedMtlArray{T},
         # we can cover the dimensions to reduce using a single group
         @metal threads=threads grid=groups partial_mapreduce_device(
             f, op, init, Val(threads), Val(Rreduce), Val(Rother),
-            Val(UInt64(length(Rother))), Val(stride), Val(shuffle), R′, A)
+            Val(UInt64(length(Rother))), Val(grain), Val(shuffle), R′, A)
     else
         # we need multiple steps to cover all values to reduce
         partial = similar(R, (size(R)..., reduce_groups))
@@ -232,7 +232,7 @@ function GPUArrays.mapreducedim!(f::F, op::OP, R::WrappedMtlArray{T},
         end
         @metal threads=threads grid=groups partial_mapreduce_device(
             f, op, init, Val(threads), Val(Rreduce), Val(Rother),
-            Val(UInt64(length(Rother))), Val(stride), Val(shuffle), partial, A)
+            Val(UInt64(length(Rother))), Val(grain), Val(shuffle), partial, A)
 
         GPUArrays.mapreducedim!(identity, op, R′, partial; init=init)
     end