Using ideas from:
* https://kieber-emmons.medium.com/optimizing-parallel-reduction-in-metal-for-apple-m1-8e8677b49b01
* https://developer.download.nvidia.com/assets/cuda/files/reduction.pdf

This replaces the old "reduce serially across chunks of input vector that don't fit in a group" loop, which doesn't seem to apply to Metal.jl since we always launch enough threadgroups to cover the entire input.
  
In a simple test this is 2-3x faster than HEAD and achieves > 130GB/s in the 1D reduction case. I'll need to experiment with the ~~stride~~ grain size parameter and document it, and test more complex cases like such as non-power-of-2 sizes or reducing only certain dimensions.

- [x] Experiment with grain size parameter and document it
- [x] Test performance of non-power-of-2 sizes
- [x] Test performance of reducing certain dimensions of N-D arrays
- [x] Check the behavior when number of threadgroups is too high
- [x] Write more tests as needed

Before:

```julia
function init(dims...)
  a = Array{Float32}(undef, dims...)
  for i in 1:length(a)
    a[i] = 1
  end
  return (a, MtlArray(a))
end

a, da = init(8192 * 8192);
b, db = init(8192, 8192);

julia> @btime sum(da)
  6.033 ms (754 allocations: 20.80 KiB)
6.7108864f7

julia> @btime sum(db)
  7.424 ms (759 allocations: 21.33 KiB)
6.7108864f7
```

After:

```julia
julia> @btime sum(da)
  2.030 ms (775 allocations: 21.20 KiB)
6.7108864f7

julia> @btime sum(db)
  3.009 ms (780 allocations: 21.72 KiB)
6.7108864f7
```

Helps with #46