@maxwindiff 2023-03-04
4 seems to be the ideal grain size based on testing in https://gist.github.com/maxwindiff/63b4e8f64c73c4b5b4ff00dd4fb79c5d

It's a bit eerie that the running times are exactly the same for all scalar data types, whether it's `UInt8`, `Float32` or `Int64`, but I don't see anything obviously wrong with the benchmarking code.