@maxwindiff 2023-03-05
I don't seem to be able to capture timing information. Do you have any idea what I'm missing? I remember seeing the timing graphs once, but I couldn't get it to work again.

```julia
julia> da = MtlArray(fill(Float32(1), 8192 * 8192));
2023-03-04 16:02:35.649 julia18[9346:290846] Metal GPU Frame Capture Enabled

julia> Metal.@profile sum(da)
[ Info: GPU frame capture saved to /Users/kichi/.julia/dev/Metal/julia_capture_1.gputrace/
6.7108864f7
```

<img width="1512" alt="Screenshot 2023-03-04 at 16 05 56" src="https://user-images.githubusercontent.com/375276/222934627-af9c6592-4770-458a-a0d6-f6ac6a428d7c.png">
<img width="1512" alt="Screenshot 2023-03-04 at 16 05 50" src="https://user-images.githubusercontent.com/375276/222934619-f2facc71-7ac9-401b-b97c-8ff4ecb1c5a9.png">
<img width="1512" alt="Screenshot 2023-03-04 at 16 06 25" src="https://user-images.githubusercontent.com/375276/222934629-4d01690d-4ab1-4d7b-b460-e0525580eff0.png">