@maxwindiff 2023-03-05
I think I know what happened, I didn't use `$<var>` in `@btime`. Need to redo the tests.