@maxwindiff 2023-03-05
Re-did the benchmarks, this time things look more reasonable: https://gist.github.com/maxwindiff/3ff4886b3e6c1d8ba9ecafb13f84eab7

A grain size of `16 / sizeof(T)` generally works well, except when reducing along certain dimensions of an n-d array. I can see why it isn't helpful to, say, read multiple values along the 2nd dimension of a 3D array. But it was surprising to see reading along the 1st dimension slowing things down... I need to investigate more.

```
sum((256, 512, 512); dims=[1])
grain=1 :  7.610 ms (272 allocations: 7.86 KiB)
grain=2 :  8.056 ms (273 allocations: 7.88 KiB)
grain=4 :  9.150 ms (273 allocations: 7.88 KiB)
grain=8 :  11.315 ms (273 allocations: 7.88 KiB)
grain=16:  20.130 ms (273 allocations: 7.88 KiB)

sum((256, 512, 512); dims=[2])
grain=1 :  9.178 ms (278 allocations: 7.95 KiB)
grain=2 :  8.769 ms (278 allocations: 7.95 KiB)
grain=4 :  9.287 ms (278 allocations: 7.95 KiB)
grain=8 :  12.515 ms (278 allocations: 7.95 KiB)
grain=16:  22.905 ms (278 allocations: 7.95 KiB)
```