@maxwindiff 2023-03-05
Made a few more changes:
* Only read multiple values if size of reduction dimension is greater than threads_per_threadgroup (otherwise we'll just starve threads of work for no good reason; this is why `sum((256,512,512); dim=1)` was slow before)
* Only read multiple values if reduction dimension is potentially contiguous in memory

Now performance is better across the board:
HEAD: https://gist.github.com/maxwindiff/4d861ae188c89ea48dd9c2986e349ca4
This PR: https://gist.github.com/maxwindiff/dcb8f3bcc926cbfc3ef55940fd35f345