@maxwindiff 2023-03-06
The reduction tests in gpuarrays seems pretty comprehensive already. This is ready for review.