@maxwindiff 2023-03-06
Actually thinking about it more, the array dimensions in `gpuarrays/reductions` are all pretty small and would not trigger multi-read. I should add a few test cases with larger array sizes.