@maxwindiff 2023-03-08
Adding a test here - https://github.com/JuliaGPU/GPUArrays.jl/pull/459
It passes on nightly, but on other platforms it had a failure when testing `Float32`. I couldn't reproduce the failure locally on 1.8 however. @maleadt do you have any suggestions on how to investigate?