@maxwindiff 2023-03-09
Done, rebased! Yep this passes the new GPUArrays tests.